Using combination of CryptoCompare and CoinMarketCap for data, removed api selection, added 7 day percent change

+++ MyCryptoMonitor/ApiData/ApiCryptoCompare.cs
@@ -7,29 +7,29 @@
         public string FROMSYMBOL { get; set; }
         public string TOSYMBOL { get; set; }
         public string FLAGS { get; set; }
-        public float PRICE { get; set; }
+        public decimal PRICE { get; set; }
         public int LASTUPDATE { get; set; }
-        public float LASTVOLUME { get; set; }
-        public float LASTVOLUMETO { get; set; }
+        public decimal LASTVOLUME { get; set; }
+        public decimal LASTVOLUMETO { get; set; }
         public string LASTTRADEID { get; set; }
-        public float VOLUMEDAY { get; set; }
-        public float VOLUMEDAYTO { get; set; }
-        public float VOLUME24HOUR { get; set; }
-        public float VOLUME24HOURTO { get; set; }
-        public float OPENDAY { get; set; }
-        public float HIGHDAY { get; set; }
-        public float LOWDAY { get; set; }
-        public float OPEN24HOUR { get; set; }
-        public float HIGH24HOUR { get; set; }
-        public float LOW24HOUR { get; set; }
+        public decimal VOLUMEDAY { get; set; }
+        public decimal VOLUMEDAYTO { get; set; }
+        public decimal VOLUME24HOUR { get; set; }
+        public decimal VOLUME24HOURTO { get; set; }
+        public decimal OPENDAY { get; set; }
+        public decimal HIGHDAY { get; set; }
+        public decimal LOWDAY { get; set; }
+        public decimal OPEN24HOUR { get; set; }
+        public decimal HIGH24HOUR { get; set; }
+        public decimal LOW24HOUR { get; set; }
         public string LASTMARKET { get; set; }
-        public float CHANGE24HOUR { get; set; }
-        public float CHANGEPCT24HOUR { get; set; }
-        public float CHANGEDAY { get; set; }
-        public float CHANGEPCTDAY { get; set; }
-        public int SUPPLY { get; set; }
-        public float MKTCAP { get; set; }
-        public float TOTALVOLUME24H { get; set; }
-        public float TOTALVOLUME24HTO { get; set; }
+        public decimal CHANGE24HOUR { get; set; }
+        public decimal CHANGEPCT24HOUR { get; set; }
+        public decimal CHANGEDAY { get; set; }
+        public decimal CHANGEPCTDAY { get; set; }
+        public long SUPPLY { get; set; }
+        public decimal MKTCAP { get; set; }
+        public decimal TOTALVOLUME24H { get; set; }
+        public decimal TOTALVOLUME24HTO { get; set; }
     }
 }
